% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gehan}
\alias{gehan}
\title{Gehan test}
\usage{
gehan(x, y, delta_x, delta_y, n1, n2, p, k = 1L, l = 1L)
}
\arguments{
\item{x}{Matrix}

\item{y}{Matrix}

\item{n1}{Integer. Set as rows in x}

\item{n2}{Integer. Set as rows in y}

\item{p}{Integer. No clue}

\item{delta.x}{Matrix}

\item{delta.y}{Matrix}
}
\value{
1x1 matrix containing a numeric
}
\description{
Performs the gehan test
}
\examples{
data <- as.matrix(wltestdata)
x <- data[1:23, c(2, 4, 6, 8)]
y <- data[24:47, c(2, 4, 6, 8)]
delta.x <- data[1:23, c(3, 5, 7, 9)]
delta.y <- data[24:47, c(3, 5, 7, 9)]
n1 <- dim(x)[1]
n2 <- dim(y)[1]
p <- dim(x)[2]
gehan(x, y, delta.x, delta.y, n1, n2, p)
}
