% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wltestdata.R
\docType{data}
\name{wltestdata}
\alias{wltestdata}
\title{Data randomly generated.}
\format{A dataframe containing 47 rows and 9 columns
\describe{
  \item{V1:V9}{Randomly generated integers}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
wltestdata
}
\description{
Data randomly generated.
}
\keyword{datasets}
