% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gehan}
\alias{gehan}
\title{Gehan test}
\usage{
gehan(x, y, delta_x, delta_y, n1, n2, p, k = 1L, l = 1L)
}
\arguments{
\item{x}{Matrix}

\item{y}{Matrix}

\item{n1}{Integer. Set as the number of rows in x}

\item{n2}{Integer. Set as the number of rows in y}

\item{p}{Integer. Set as the number of columns in x and y}

\item{delta.x}{Matrix}

\item{delta.y}{Matrix}
}
\value{
1x1 matrix containing a numeric
}
\description{
Computes the multivariate Gehan test statistic.
}
