% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finished_permutations.R
\name{perm_mvlogrank}
\alias{perm_mvlogrank}
\title{Permutations of mvlogrank test}
\usage{
perm_mvlogrank(B, z, delta.z, n1, n2, p, mv.test)
}
\arguments{
\item{B}{Integer specifying the number of permutations to perform}

\item{z}{A matrix which combines the x and y matricies}

\item{delta.z}{A matrix which combines the delta.x and delta.y matricies}

\item{n1}{Integer. Should be specified as number of rows in the x matrix}

\item{n2}{Integer. Shouled be specified as number of rows in the y matrix}

\item{p}{Integer. No clue what it does}

\item{g.test}{Numeric from gehan}
}
\value{
Numeric
}
\description{
Performs permutations of the mvlogrank test through random sub sampling
}
\examples{
See ?MultSurvTests

}
